fixing lifecycle validation

+++ validator/expression_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestCallExpressionValid(t *testing.T) {
 	scope, _ := parser.ParseString(`
-        func call(a, b int) int {
+        func call(a, b int8) int8 {
             if a == 0 or b == 0 {
                 return 0
             }
